backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
pgvector==0.4.2

# OCR - using headless OpenCV for Docker
# Pin to v2.x for stable API and numpy 1.x for opencv compatibility
numpy<2.0
paddlepaddle==2.6.2
paddleocr==2.7.3
opencv-python-headless>=4.8.0

# LLM (optional)
openai==1.109.1
langchain==1.2.2
langchain-openai==1.1.7
langsmith==0.6.1

# Utilities
python-dotenv==1.0.0
pydantic==2.12.5
pydantic-settings==2.12.0
Pillow==10.2.0
PyYAML==6.0.1
jsonschema==4.21.1
pymupdf==1.26.7
pdf2image==1.17.0
pdfplumber==0.10.3
openpyxl==3.1.2

# Rate limiting
slowapi==0.1.9

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
